DO NOT MERGE: Build static libevent for the target.
libevent is already built as a static library in master since it was
converted to a .bp file. This patch builds it as a static library using
the legacy .mk file in this branch.

Bug: 27178350
Change-Id: Ic677b45a5ba806f688d766265f746be33dde7409

Cherry-pick needed on this branch but not on the downstream ones.
(cherry picked from commit a5dd84c71082d993bd8273d674daf89e4e9853a7)

Change-Id: Ied4bae6a46a48929b142cab8dfbcd325cf09ff18